Make the network caches a separate package
This gives more structure to the code and avoid circular dependencies.
