+++ pkg/virt-launcher/virtwrap/network/cache/fake/infocache.go
@@ -1,15 +1,15 @@
 package fake
 
 import (
-	"fmt"
+	"os"
 	"sync"
 
 	"k8s.io/apimachinery/pkg/types"
+
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/network/cache"
 
 	v1 "kubevirt.io/client-go/api/v1"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/api"
-	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/network"
 )
 
 func NewFakeInMemoryNetworkCacheFactory() cache.InterfaceCacheFactory {
@@ -32,7 +32,10 @@ func (f *fakeInterfaceCacheFactory) CacheForVMI(vmi *v1.VirtualMachineInstance) 
 	if store, exists := f.vmiCacheStores[vmi.UID]; exists {
 		return store
 	}
-	f.vmiCacheStores[vmi.UID] = &fakePodInterfaceCacheStore{lock: &sync.Mutex{}}
+	f.vmiCacheStores[vmi.UID] = &fakePodInterfaceCacheStore{
+		store: map[string]*cache.PodCacheInterface{},
+		lock:  &sync.Mutex{},
+	}
 	return f.vmiCacheStores[vmi.UID]
 }
 
@@ -42,25 +45,28 @@ func (f *fakeInterfaceCacheFactory) CacheForPID(pid string) cache.DomainInterfac
 	if store, exists := f.domainCacheStores[pid]; exists {
 		return store
 	}
-	f.domainCacheStores[pid] = &fakeDomainInterfaceStore{lock: &sync.Mutex{}}
+	f.domainCacheStores[pid] = &fakeDomainInterfaceStore{
+		store: map[string]*api.Interface{},
+		lock:  &sync.Mutex{},
+	}
 	return f.domainCacheStores[pid]
 }
 
 type fakePodInterfaceCacheStore struct {
 	lock  *sync.Mutex
-	store map[string]*network.PodCacheInterface
+	store map[string]*cache.PodCacheInterface
 }
 
-func (f *fakePodInterfaceCacheStore) Read(iface string) (*network.PodCacheInterface, error) {
+func (f *fakePodInterfaceCacheStore) Read(iface string) (*cache.PodCacheInterface, error) {
 	f.lock.Lock()
 	defer f.lock.Unlock()
 	if val, exists := f.store[iface]; exists {
 		return val, nil
 	}
-	return nil, fmt.Errorf("no cache file for interface %s", iface)
+	return nil, os.ErrNotExist
 }
 
-func (f *fakePodInterfaceCacheStore) Write(iface string, cacheInterface *network.PodCacheInterface) error {
+func (f *fakePodInterfaceCacheStore) Write(iface string, cacheInterface *cache.PodCacheInterface) error {
 	f.lock.Lock()
 	defer f.lock.Unlock()
 	f.store[iface] = cacheInterface
@@ -70,7 +76,7 @@ func (f *fakePodInterfaceCacheStore) Write(iface string, cacheInterface *network
 func (f *fakePodInterfaceCacheStore) Remove() error {
 	f.lock.Lock()
 	defer f.lock.Unlock()
-	f.store = map[string]*network.PodCacheInterface{}
+	f.store = map[string]*cache.PodCacheInterface{}
 	return nil
 }
 
@@ -85,7 +91,7 @@ func (f *fakeDomainInterfaceStore) Read(iface string) (*api.Interface, error) {
 	if val, exists := f.store[iface]; exists {
 		return val, nil
 	}
-	return nil, fmt.Errorf("no cache file for interface %s", iface)
+	return nil, os.ErrNotExist
 }
 
 func (f *fakeDomainInterfaceStore) Write(iface string, cacheInterface *api.Interface) error {
